+++ engine/base/request-response/runtime/src/main/scala/pl/touk/nussknacker/engine/requestresponse/deployment/ProcessInterpreters.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.engine.requestresponse.deployment
+
+import pl.touk.nussknacker.engine.requestresponse.RequestResponseEngine
+
+trait ProcessInterpreters {
+
+  def getInterpreterByPath(path: String): Option[RequestResponseEngine.RequestResponseScenarioInterpreter]
+
+}