+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/api/ResponseEncoder.scala
@@ -1,9 +1,0 @@
-package pl.touk.nussknacker.engine.standalone.api
-
-import io.circe.Json
-
-trait ResponseEncoder[-Input] {
-
-  def toJsonResponse(input: Input, result: List[Any]): Json
-
-}