+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/api/openapi/OpenApiSourceDefinition.scala
@@ -1,5 +1,0 @@
-package pl.touk.nussknacker.engine.standalone.api.openapi
-
-import io.circe.Json
-
-case class OpenApiSourceDefinition(definition: Json, description: String, tags: List[String])